refactor/テストコードのモジュール化によるリファクタリング > user_controllerのテスト
